refac: create projects table component
